net/mlx5: Update max size for nvme_queue_size
According to NVMe specification, the maximal NVMe device queue
size is 2^16 elements. In our NVMEoF target offload solution we
dedicate only 16 bits to this field so use 0 as 2^16 to support
the maximal value for an NVMe device.

+++ include/linux/mlx5/srq.h
@@ -59,7 +59,7 @@ struct mlx5_nvmf_attr {
 	u8				staging_buffer_log_page_size;
 	u16				staging_buffer_number_of_pages;
 	u8				staging_buffer_page_offset;
-	u16				nvme_queue_size;
+	u32				nvme_queue_size;
 	u64				*staging_buffer_pas;
 };
